Added most up to date version of spider
This spider is an edited version of the one that scraped the site. The one that scraped the site had some issues we found while exploring the data so I updated it in case there were ever a reason to scrape the entire site again. Another spider was run to re-scrape only the threads containing problem posts.